+++ src/main/java/com/imsweb/naaccrxml/PatientXmlReader.java
@@ -300,13 +300,8 @@ public class PatientXmlReader implements PatientReader {
 
             // handle root extensions
             while (!isPatientTag(_reader.getNodeName()) && !isRootTag(_reader.getNodeName())) {
-                if (!Boolean.TRUE.equals(options.getIgnoreExtensions())) {
-                    int lineNumber = conf.getParser().getLineNumber();
-                    Object extension = conf.getXstream().unmarshal(_reader);
-                    if (extension instanceof NaaccrXmlExtension)
-                        ((NaaccrXmlExtension)extension).setStartLineNumber(lineNumber);
-                    _rootData.addExtension(extension);
-                }
+                if (!Boolean.TRUE.equals(options.getIgnoreExtensions()))
+                    _rootData.addExtension(conf.getXstream().unmarshal(_reader));
                 _reader.moveUp();
                 if (_reader.hasMoreChildren())
                     _reader.moveDown();